refactor: Header에서 모달 컴포넌트 분리하기
